+++ demos/chinook/src/main/java/is/codion/framework/demos/chinook/ui/CustomerTablePanel.java
@@ -18,7 +18,6 @@
  */
 package is.codion.framework.demos.chinook.ui;
 
-import is.codion.common.db.report.ReportException;
 import is.codion.framework.demos.chinook.domain.Chinook.Customer;
 import is.codion.framework.domain.entity.Entity;
 import is.codion.swing.common.ui.control.Control;
@@ -66,7 +65,7 @@ public final class CustomerTablePanel extends EntityTablePanel {
 						.execute();
 	}
 
-	private JasperPrint fillCustomerReport() throws ReportException {
+	private JasperPrint fillCustomerReport() {
 		Collection<Long> customerIDs = Entity.values(Customer.ID,
 						tableModel().selection().items().get());
 		Map<String, Object> reportParameters = new HashMap<>();