+++ changelog.md
@@ -4,6 +4,7 @@ Codion Change Log
 ## 0.18.21
 ### is.codion.common.db
 - DatabaseException now extends RuntimeException, no longer checked.
+- ReportException now extends RuntimeException, no longer checked.
 ### is.codion.swing.common.ui
 - ExceptionDialogBuilder.unwrap() added.
 ### is.codion.framework.domain